+++ test/test_HPE3ParClient_FilePersona_Mock.py
@@ -18,7 +18,7 @@ import mock
 import pprint
 
 from testconfig import config
-import HPE3ParClient_base as hpe3parbase
+from test import HPE3ParClient_base as hpe3parbase
 
 from hpe3parclient import exceptions
 from hpe3parclient import file_client