+++ test/test_HPE3ParClient_system.py
@@ -16,7 +16,7 @@
 
 from testconfig import config
 import unittest
-import HPE3ParClient_base as hpe3parbase
+from test import HPE3ParClient_base as hpe3parbase
 
 from hpe3parclient import exceptions